The widest integer there is, and which box decides how wide that is
C-04, and with it the last of the nineteen the audit found.

**[cpp.cond] evaluates a condition in the widest signed integer there is**, and
five levels of this evaluator worked in `long`. A `long` is 64 bits where gcc
and clang build this compiler and 32 where cl does, so a condition holding a
value wider than 32 bits answered differently depending on which of the three
boxes had built the preprocessor reading it - `#if 0x300000002 & 0xFFFFFFFF`
among them. The same source, two answers, and no test could see it from one
machine. The bit-field widths went the same way as the array length before
them, for the same reason.

**And the arithmetic wraps rather than overflowing.** This evaluator runs
inside a compiler, so undefined behaviour on overflow would mean the compiler
itself, and three shapes were reachable from a `#if` line: negating the most
negative value, the one division that overflows, and a shift count outside 0
to 63. The first two are defined through the unsigned type, which is what a
preprocessor is expected to do; the third is refused by name, because a count
of 100 is a mistake and not a large shift.

Mac 223/348/117/26, Linux 223/348, Windows 220 cases and 93 names vs cl - and
the Windows box is the one this was about, being the box where a long is 32
bits.

All nineteen findings are closed. The suite went from 201 run cases to 223 and
from 292 emissions to 348 along the way.

+++ src/parser/ParserType.cpp
@@ -501,7 +501,7 @@ const Type *Parser::structOrUnionSpecifier(Kind kind, bool isClass) {
             if (peek().is(":")) {
                 std::size_t cpos = peek().pos;
                 at_++;
-                long w = constantExpression("a bit-field width");
+                long long w = constantExpression("a bit-field width");
                 if (!base->isInteger())
                     src_.fail(cpos, "a bit-field must have an integer type, not '" +
                                     base->describe() + "'");
@@ -602,7 +602,7 @@ const Type *Parser::structOrUnionSpecifier(Kind kind, bool isClass) {
             if (peek().is(":")) {
                 std::size_t cpos = peek().pos;
                 at_++;
-                long w = constantExpression("a bit-field width");
+                long long w = constantExpression("a bit-field width");
                 if (!d.type->isInteger())
                     src_.fail(d.pos, "a bit-field must have an integer type, not '" +
                                      d.type->describe() + "'");
